+++ src/main/java/com/team3/prj/service/QnaServiceImpl.java
@@ -20,7 +20,7 @@ public class QnaServiceImpl implements QnaService {
 	}
 
 	@Override
-	public Integer insert(QnaVO vo) {
+	public Long insert(QnaVO vo) {
 		return dao.qnaInsert(vo);
 	}
 }
